Updated the module displayDate.js
