+++ index.html
@@ -53,7 +53,6 @@
     <footer class="copyright">
         <p>Copyright...</p>
     </footer>
-    <!-- <script src="https://cdn.jsdelivr.net/npm/luxon@2.1.1/build/global/luxon.min.js" type="module"></script> -->
     <script src="index.js" type="module"></script>
 </body>
